refactor(tests): rename the tf_cli_config fixture and the cli module
tf_cli_config becomes tool_config and fixtures/cli.py becomes
fixtures/tool.py, so the whole seam is named after the --tool flag.
